@@ -165,17 +165,17 @@
     "wreck": "6.2.0"
   },
   "devDependencies": {
-    "@elastic/eslint-config-kibana": "0.0.3",
+    "@elastic/eslint-config-kibana": "0.2.1",
     "angular-mocks": "1.4.7",
     "auto-release-sinon": "1.0.3",
-    "babel-eslint": "4.1.8",
+    "babel-eslint": "6.1.2",
     "chai": "3.5.0",
     "cheerio": "0.22.0",
     "chokidar": "1.6.0",
     "chromedriver": "2.24.1",
     "elasticdump": "2.1.1",
-    "eslint": "1.10.3",
-    "eslint-plugin-mocha": "1.1.0",
+    "eslint": "3.11.1",
+    "eslint-plugin-mocha": "4.7.0",
     "event-stream": "3.3.2",
     "expect.js": "0.3.1",
     "faker": "1.1.0",
@@ -189,7 +189,6 @@
     "grunt-karma": "2.0.0",
     "grunt-run": "0.6.0",
     "grunt-simple-mocha": "0.4.0",
-    "gruntify-eslint": "1.0.1",
     "gulp-sourcemaps": "1.7.3",
     "handlebars": "4.0.5",
     "husky": "0.8.1",
