@@ -1,14 +1,14 @@
-const rimraf = promisify(require('rimraf'));
-const mkdirp = promisify(require('mkdirp'));
-const unlink = promisify(require('fs').unlink);
-const readdir = promisify(require('fs').readdir);
-
 import UiBundle from './ui_bundle';
 import appEntryTemplate from './app_entry_template';
 import { readFileSync as readSync } from 'fs';
 import { pull, transform, pluck } from 'lodash';
 import { promisify } from 'bluebird';
 import { makeRe } from 'minimatch';
+
+const rimraf = promisify(require('rimraf'));
+const mkdirp = promisify(require('mkdirp'));
+const unlink = promisify(require('fs').unlink);
+const readdir = promisify(require('fs').readdir);
 
 class UiBundleCollection {
   constructor(bundlerEnv, filter) {
