@@ -73,8 +73,9 @@ bdd.describe('visualize app', function describeIndexTests() {
 
     bdd.it('should show correct data, take screenshot', function () {
       var chartHeight = 0;
-      var expectedChartData = [ '0 2,088', '2,000 2,748', '4,000 2,707', '6,000 2,876',
-      '8,000 2,863', '10,000 147', '12,000 148', '14,000 129', '16,000 161', '18,000 137'
+      var expectedChartData = [
+        '0 2,088', '2,000 2,748', '4,000 2,707', '6,000 2,876',
+        '8,000 2,863', '10,000 147', '12,000 148', '14,000 129', '16,000 161', '18,000 137'
       ];
 
       return PageObjects.common.try(function () {
